+++ lab 1/lab 1/Triangle.h
@@ -10,9 +10,9 @@ class Triangle: public Polygon
 public:
 	Triangle();
 	Triangle(pair<double, double> p[3]);
-	Triangle(int x1, int y1, int x2, int y2, int x3, int y3);
+	Triangle(double x1, double y1, double x2, double y2, double x3, double y3);
 	void set_points(pair<double, double> p[3]);
-	void set_points(int x1, int y1, int x2, int y2, int x3, int y3);
+	void set_points(double x1, double y1, double x2, double y2, double x3, double y3);
 	double square();
 	double perimetr();
 	string regular();
